dto fromDogArray
