+++ src/core/use_cases/dogs/get.all.breeds.ts
@@ -1,7 +1,6 @@
 import { DogsService } from '../../../services/dogs/dogs.service';
 import { Injectable } from '@nestjs/common';
-import { DogsOutputDto } from 'src/core/dto/dogs/dogs.output.dto';
-import { Dog } from 'src/core/entities/dog';
+import { DogsOutputDto, fromDogArray } from 'src/core/dto/dogs/dogs.output.dto';
 
 @Injectable()
 export class GetAllBreeds {
@@ -9,21 +8,6 @@ export class GetAllBreeds {
 
   public async call(): Promise<DogsOutputDto> {
     const dogs = await this.dogService.allBreeds();
-    return this.toListandNumber(dogs);
-  }
-
-  private toListandNumber(dogs: Dog[]) {
-    let breedNamesList: string[] = [];
-
-    Object.keys(dogs).forEach(Element => {
-      breedNamesList.push(dogs[Element].breed);
-    });
-
-    const adapted = {
-      list: breedNamesList,
-      total: breedNamesList.length,
-    };
-
-    return adapted;
+    return fromDogArray(dogs);
   }
 }
